Python/Coding/Development/Django/supabase_django_project/authentication/urls.py
# authentication/urls.py
from django.urls import path, include  # Added the include import
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter()
router.register('users', views.UserViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('register/', views.register_user, name='register'),
    path('login/', views.login_user, name='login'),
    path('logout/', views.logout_user, name='logout'),
]